requirements.txt
flask
openai-whisper
deep-translator
torch
ffmpeg-python
